Extract tracking and creating file transfers to ILayoutServices
This refactoring will allow more parts of the pipeline to add files
to be transferred during layout and tracked to participate in
MSBuild up to date checks and cleaning.
