+++ src/wix/WixToolset.Core/WixToolsetServiceProvider.cs
@@ -23,6 +23,7 @@ namespace WixToolset.Core
             this.AddService((provider, singletons) => AddSingleton<ISymbolDefinitionCreator>(singletons, new SymbolDefinitionCreator(provider)));
             this.AddService((provider, singletons) => AddSingleton<IParseHelper>(singletons, new ParseHelper(provider)));
             this.AddService((provider, singletons) => AddSingleton<IPreprocessHelper>(singletons, new PreprocessHelper(provider)));
+            this.AddService((provider, singletons) => AddSingleton<ILayoutServices>(singletons, new LayoutServices(provider)));
             this.AddService((provider, singletons) => AddSingleton<IBackendHelper>(singletons, new BackendHelper(provider)));
             this.AddService((provider, singletons) => AddSingleton<IPathResolver>(singletons, new PathResolver()));
             this.AddService((provider, singletons) => AddSingleton<IWixBranding>(singletons, new WixBranding()));
